Comments for blog posts
PostCommentForm is displayed at bottom of blog posts for users who have
filled their profile. The form is submitted as AJAX request. The list of
all comments show up above the form on page refresh.

List of all comments are requested asynchronously and updated once
in a while without reloading the page.